Filter out `-static` option on MacOS X
This is to fix gprbuild 24 on Mac OS. The error is

> ld: library not found for -lcrt0.o

The gprbuild call `gcc -static ...` during linking a main, but
MacOS X was never able to do [static linking][1].

The gprbuild reads `-static` option from `*.bexch` file.
The `*.bexch` file is created by `gprbind` by reading
`b__<main>.adb` generated by `gnatbind`.
The `gnatbind` writes `-static` to `b__<main>.adb` for ages,
but gprbuild 22 filtered this option.
The filtering was removed in [the commit][2].

[1]: https://stackoverflow.com/questions/3801011/ld-library-not-found-for-lcrt0-o-on-osx-10-6-with-gcc-clang-static-flag
[2]: https://github.com/AdaCore/gprbuild/commit/48ea3c19d4f5552109642f339d22c58308970c42