Build GCC 14.1 on aarch64-apple-darwin.
Use Xcode 15.3, because (at this time) the macos-14 runner defaults to
Xcode 15.0.1, which has the linking bug. 15.3 is the latest version
available in Command Line Tools; Xcode has 15.4.

Use GCC 13.3.0 as the base compiler, because (on Github) 13.2 links one
of the tools generated during the compiler build in such a way that it
won't run.

Update GNAT's aarch64 source to the gcc-14.1-darwin-r1 release.

Don't configure GCC with --without-build-config (the reason for
using it was GCC PR 100340, which was fixed in 2022; using it caused
failures in bootstrap stage 2/3 comparison).

  * .github/workflows/macos_arm.yml (GNAT macOS): use Xcode 15.3.
    (SPARK): likewise.
  * specs/base_gcc.anod (tarball): directly name the 13.3.0 build
      to be used.
    (url): directly name the location of the 13.3.0 build to be used.
  * specs/gcc.anod (source_pkg_build): update to the gcc-14.1-darwin-r1
      release.
    (configure_args): don't use --without-build-config.